Fixed bm to gcode algorithm bugs
Now it will create the correct gcode for cylindar or rectangular when there are special cases such as pixes drawn tll the last column, and different combinations of of engraving/notengraving when moving to next row. Added a function that will just straight up send a string that it is passed to the stm32f0.

TODO Change the SPI sending functions to be UART